requirements.txt
pandas==2.2.3
torch==2.4.1
langchain==0.3.0
langchain_community==0.3.0
langchain-huggingface==0.1.0
langchain_core==0.3.5
huggingface-hub==0.24.7
sentence_transformers==3.1.1 
qdrant-client==1.11.2
accelerate==0.34.2